# api/docs/swagger.yaml
definitions:
  models.BadRequestError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.CreateResponse:
    properties:
      id:
        type: string
    type: object
  models.CreateUpdateResponsibleRequest:
    properties:
      comment:
        type: string
      created_at:
        type: integer
      name_of_step:
        type: string
      organization:
        type: string
      role:
        type: string
      updated_at:
        type: integer
    type: object
  models.GetAllResponsiblesResponse:
    properties:
      count:
        type: integer
      responsibles:
        items:
          $ref: '#/definitions/models.Responsible'
        type: array
    type: object
  models.GetResponsibleResponse:
    properties:
      responsible:
        $ref: '#/definitions/models.Responsible'
    type: object
  models.InternalServerError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.Responsible:
    properties:
      comment:
        type: string
      created_at:
        type: integer
      id:
        type: string
      name_of_step:
        type: string
      organization:
        type: string
      role:
        type: string
      updated_at:
        type: integer
    type: object
info:
  contact: {}
paths:
  /v1/responsible:
    get:
      consumes:
      - application/json
      description: API for getting all responsibles
      parameters:
      - description: name
        in: path
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllResponsiblesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Get All responsibles
      tags:
      - responsible
    post:
      consumes:
      - application/json
      description: API for creating responsible
      parameters:
      - description: responsible
        in: body
        name: responsible
        required: true
        schema:
          $ref: '#/definitions/models.CreateUpdateResponsibleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Create responsible
      tags:
      - responsible
  /v1/responsible/{responsible_id}:
    get:
      consumes:
      - application/json
      description: API for getting a responsible
      parameters:
      - description: responsible_id
        in: path
        name: responsible_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetResponsibleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Get responsible
      tags:
      - responsible
    put:
      consumes:
      - application/json
      description: API for creating responsible
      parameters:
      - description: responsible_id
        in: path
        name: responsible_id
        required: true
        type: string
      - description: responsible
        in: body
        name: responsible
        required: true
        schema:
          $ref: '#/definitions/models.CreateUpdateResponsibleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Update responsible
      tags:
      - responsible
swagger: "2.0"